+++ Control/remote/remote.c
@@ -1,19 +1,19 @@
 #include "remote.h"
 
-int32_t remote_pitch_angle;
-int32_t remote_roll_angle;
-int32_t remote_yaw_angle;
+float remote_pitch_angle;
+float remote_roll_angle;
+float remote_yaw_angle;
 
-int32_t remote_pitch_w;
-int32_t remote_roll_w;
-int32_t remote_yaw_w;
+float remote_pitch_w;
+float remote_roll_w;
+float remote_yaw_w;
 
-int32_t remote_h;
+float remote_h;
 
-signed int remote_x;
-signed int remote_y;
-signed int remote_z;
-signed int remote_yaw;
+float remote_x;
+float remote_y;
+float remote_z;
+float remote_yaw;
 
 void Updata_set(void)
 {
